+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this project are documented in this file.
 
 - univesal node: player.get.buttonlockstate, player.set.buttonlockstate
 
+- universal node: group.play.httpstream may use msg.info, msg.artUri to display title, use cover
+
 ### [5.0.0] 2021-03-13
 
 #### Changed